Added a loading spinner and fixed some bugs with sizing and removal

+++ client/views/ToDoList.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import axios from 'axios'
 import moment from 'moment'
+import Loader from 'react-loaders'
 import { ToDoAdd } from '../components/ToDoAdd'
 import { ToDoTable } from '../components/ToDoTable'
 import PropTypes from 'prop-types'
@@ -12,6 +13,7 @@ class ToDoList extends Component {
   constructor (props) {
     super(props)
     this.state = {
+      loading: false,
       editId: '',
       list: [],
       todo: {
@@ -28,6 +30,7 @@ class ToDoList extends Component {
     this._deleteToDo = this._deleteToDo.bind(this)
     this._editToDo = this._editToDo.bind(this)
     this._saveChanges = this._saveChanges.bind(this)
+    this._cancelChanges = this._cancelChanges.bind(this)
   }
   componentWillMount () {
     axios.get('/todo/').then(response => {
@@ -40,7 +43,7 @@ class ToDoList extends Component {
   _addToDo () {
     if (this.state.todo.todo !== '' && this.state.todo.due !== '') {
       let newToDo = Object.assign({}, this.state.todo, { created: Date.now(), completed: false }, {due: moment(this.state.todo.due, 'YYYY-MM-DD').format()})
-      this.setState({ todo: newToDo })
+      this.setState({ todo: newToDo, loading: true })
       delete newToDo._id
 
       axios.post('/todo/', newToDo).then(response => {
@@ -52,7 +55,9 @@ class ToDoList extends Component {
           due: '',
           created: '',
           completed: false
-        }})
+        },
+        loading: false
+        })
       }).catch(err => {
         console.log(err)
       })
@@ -64,27 +69,45 @@ class ToDoList extends Component {
     this.setState({todo: stateObj})
   }
   _setComplete (toDo) {
+    this.setState({ loading: true })
     toDo.completed = !toDo.completed
     axios.put('/todo/' + toDo._id, toDo).then(response => {
       this.props.updateAListValue(response.data)
       this.props.updateToDo(response.data)
+      this.setState({ loading: false })
     }).catch(err => {
       console.log(err)
     })
   }
   _deleteToDo (toDo) {
-    axios.delete('/todo/' + toDo._id).then(response => {
-      this.props.removeToDo(response.data)
-    }).catch(err => {
-      console.log(err)
-    })
+    this.setState({ loading: true })
+    if (toDo) {
+      axios.delete('/todo/' + toDo._id).then(response => {
+        this.props.removeToDo(response.data)
+        this.setState({ loading: false })
+      }).catch(err => {
+        console.log(err)
+      })
+    }
   }
   _editToDo (todo) {
-    let todoObj = Object.assign({}, this.state.todo, todo)
-    delete todoObj._v
-    this.setState({ todo: todoObj, editId: todo._id })
+    if (todo) {
+      let todoObj = Object.assign({}, this.state.todo, todo)
+      delete todoObj._v
+      this.setState({ todo: todoObj })
+    }
+  }
+  _cancelChanges (todo) {
+    this.setState({todo: {
+      _id: '',
+      todo: '',
+      due: '',
+      created: '',
+      completed: false
+    }})
   }
   _saveChanges () {
+    this.setState({ loading: true })
     axios.put('/todo/' + this.state.todo._id, this.state.todo).then(response => {
       this.props.updateAListValue(response.data)
       this.setState({todo: {
@@ -94,7 +117,8 @@ class ToDoList extends Component {
         created: '',
         completed: false
       },
-      editId: ''})
+      loading: false
+      })
     }).catch(err => {
       console.log(err)
     })
@@ -106,14 +130,18 @@ class ToDoList extends Component {
           <h1 style={{ fontWeight: 200, color: 'white' }}>To Do list</h1>
         </div>
         <ToDoAdd _addToDo={this._addToDo} _updateValue={this._updateValue}/>
-        <ToDoTable
-          toDoList={this.props.ListState.list}
-          _setComplete={this._setComplete}
-          _deleteToDo={this._deleteToDo}
-          _editToDo={this._editToDo}
-          _saveChanges={this._saveChanges}
-          _updateValue={this._updateValue}
-          editData={this.state.todo}/>
+        { this.state.loading
+          ? <Loader type="line-scale" active />
+          : <ToDoTable
+            toDoList={this.props.ListState.list}
+            _setComplete={this._setComplete}
+            _deleteToDo={this._deleteToDo}
+            _editToDo={this._editToDo}
+            _saveChanges={this._saveChanges}
+            _cancelChanges={this._cancelChanges}
+            _updateValue={this._updateValue}
+            editData={this.state.todo}/>
+        }
       </div>
     )
   }
